make it work with ubuntu 14.04

+++ prompt/install_launch.sh
@@ -1,5 +1,1 @@
-if [ -f "install_prompt2.sh" ] then rm "install_prompt2.sh" fi
-wget --quiet "https://raw.githubusercontent.com/maassql/bash/master/prompt/install_prompt2.sh"
-bash "install_prompt2.sh"
-rm "install_prompt2.sh"
-source "${HOME}/.jeff_maass_prompt/jmaass_prompt.sh"
+if [ -f "install_prompt2.sh" ] ; then rm "install_prompt2.sh"; fi; wget --quiet "https://raw.githubusercontent.com/maassql/bash/master/prompt/install_prompt2.sh";bash "install_prompt2.sh";rm "install_prompt2.sh";source "${HOME}/.jeff_maass_prompt/jmaass_prompt.sh";
